requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
youtube-transcript-api==1.2.3
jinja2==3.1.3
python-docx==1.1.0
